video: warn if timestamp correction introduces drift
We have code which attempts to fix jittering timestamps due to rounding
to milliseconds in containers like mkv. This can introduce a major drift
of the muxer wrote a wrong framerate that's close to the actual
framerate.

If this happens, print a warning.

We could just ignore it, as display sync changes the playback speed in a
similar way anyway. But currently, we don't compensate for the drift
correctly, so it ends up dropping or repeating extra frames. Audio is
essentially played at the wrong speed.